requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
